Refactor tests & stories & update snapshots
