+++ src/app/components/Figure/Image/helpers/fixtureData.js
